rename : #129 Header -> Navigator로 컴포넌트 이름 수정
