+++ project_frontend/src/components/navigator/navigatorStyledComponents.js
